+++ examples/expo-relay/.wundergraph/wundergraph.server.ts
@@ -1,0 +1,12 @@
+import { configureWunderGraphServer } from '@wundergraph/sdk/server';
+
+export default configureWunderGraphServer(() => ({
+	hooks: {
+		queries: {
+			Countries: {
+				preResolve: async ({ operations }) => {},
+			},
+		},
+		mutations: {},
+	},
+}));