+++ packages/sdk/src/operations/operations.ts
@@ -1,25 +1,14 @@
 import { z } from 'zod';
 import * as fs from 'fs';
+import type { ORM } from '@wundergraph/orm';
 import type { BaseRequestContext, InternalClient, OperationsClient, WunderGraphUser } from '../server';
 import { OperationError } from '../client';
-import {
-	LiveQueryConfiguration,
-	QueryCacheConfiguration as OperationQueryCacheConfiguration,
-} from '../configure/operations';
-export type { LiveQueryConfiguration } from '../configure/operations';
-
-export type QueryCacheConfiguration = Omit<OperationQueryCacheConfiguration, 'enable'>;
+import { LiveQueryConfiguration, QueryCacheConfiguration } from '../configure/operations';
+export type { LiveQueryConfiguration, QueryCacheConfiguration } from '../configure/operations';
 
 const disabledLiveQueryConfiguration = {
 	enable: false,
 	pollingIntervalSeconds: 0,
-};
-
-const disabledCacheConfiguration = {
-	enable: false,
-	public: false,
-	maxAge: 0,
-	staleWhileRevalidate: 0,
 };
 
 export type SubscriptionHandler<
@@ -30,13 +19,14 @@ export type SubscriptionHandler<
 	UserRole extends string,
 	CustomClaims extends {},
 	InternalOperationsClient extends OperationsClient,
+	TypedORM,
 	CustomContext
 > = ZodResponse extends z.ZodObject<any>
 	? (
-			ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>
+			ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>
 	  ) => AsyncGenerator<z.infer<ZodResponse>>
 	: (
-			ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>
+			ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>
 	  ) => AsyncGenerator<InferredResponse>;
 
 export type OperationTypes = 'query' | 'mutation' | 'subscription';
@@ -47,10 +37,12 @@ interface _HandlerContext<
 	Role extends string,
 	CustomClaims extends {},
 	InternalOperationsClient extends OperationsClient,
+	TypedORM,
 	CustomContext
 > extends BaseRequestContext<WunderGraphUser<Role, CustomClaims>, IC, OperationsClient, CustomContext> {
 	input: Input extends {} ? Input : never;
 	operations: Omit<InternalOperationsClient, 'cancelSubscriptions'>;
+	graph: TypedORM;
 }
 
 export type HandlerContext<
@@ -59,10 +51,11 @@ export type HandlerContext<
 	Role extends string,
 	CustomClaims extends {},
 	InternalOperationsClient extends OperationsClient,
+	TypedORM,
 	CustomContext
 > = Input extends z.ZodObject<any>
-	? _HandlerContext<z.infer<Input>, IC, Role, CustomClaims, InternalOperationsClient, CustomContext>
-	: Omit<_HandlerContext<never, IC, Role, CustomClaims, InternalOperationsClient, CustomContext>, 'input'>;
+	? _HandlerContext<z.infer<Input>, IC, Role, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>
+	: Omit<_HandlerContext<never, IC, Role, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>, 'input'>;
 
 export interface BaseOperationConfiguration<UserRole extends string> {
 	requireAuthentication?: boolean;
@@ -81,6 +74,7 @@ const createQuery =
 		UserRole extends string,
 		CustomClaims extends {},
 		InternalOperationsClient extends OperationsClient,
+		TypedORM,
 		TCustomContext
 	>() =>
 	<Input extends z.ZodObject<any> = any, InferredResponse = unknown, ZodResponse = unknown>({
@@ -99,10 +93,10 @@ const createQuery =
 		response?: ZodResponse;
 		handler: ZodResponse extends z.ZodObject<any>
 			? (
-					ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TCustomContext>
+					ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, TCustomContext>
 			  ) => Promise<z.infer<ZodResponse>>
 			: (
-					ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TCustomContext>
+					ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, TCustomContext>
 			  ) => Promise<InferredResponse>;
 		live?: LiveQueryConfiguration;
 		cache?: QueryCacheConfiguration;
@@ -115,6 +109,7 @@ const createQuery =
 		UserRole,
 		CustomClaims,
 		InternalOperationsClient,
+		TypedORM,
 		TCustomContext
 	> => {
 		return {
@@ -131,11 +126,7 @@ const createQuery =
 				requireMatchAny: rbac?.requireMatchAny || [],
 			},
 			errors,
-			cache: {
-				...disabledCacheConfiguration,
-				...cache,
-				enable: cache !== undefined,
-			},
+			cache,
 			liveQuery: {
 				enable: live?.enable || true,
 				pollingIntervalSeconds: live?.pollingIntervalSeconds || 5,
@@ -149,6 +140,7 @@ const createMutation =
 		UserRole extends string,
 		CustomClaims extends {},
 		InternalOperationsClient extends OperationsClient,
+		TypedORM,
 		CustomContext
 	>() =>
 	<Input extends z.ZodObject<any> = any, InferredResponse = unknown, ZodResponse = unknown>({
@@ -165,10 +157,10 @@ const createMutation =
 		response?: ZodResponse;
 		handler: ZodResponse extends z.ZodObject<any>
 			? (
-					ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>
+					ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>
 			  ) => Promise<z.infer<ZodResponse>>
 			: (
-					ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>
+					ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>
 			  ) => Promise<InferredResponse>;
 	} & BaseOperationConfiguration<UserRole>): NodeJSOperation<
 		z.infer<Input>,
@@ -179,6 +171,7 @@ const createMutation =
 		UserRole,
 		CustomClaims,
 		InternalOperationsClient,
+		TypedORM,
 		CustomContext
 	> => {
 		return {
@@ -196,7 +189,6 @@ const createMutation =
 			},
 			errors,
 			liveQuery: disabledLiveQueryConfiguration,
-			cache: disabledCacheConfiguration,
 		};
 	};
 
@@ -206,6 +198,7 @@ const createSubscription =
 		UserRole extends string,
 		CustomClaims extends {},
 		InternalOperationsClient extends OperationsClient,
+		TypedORM,
 		CustomContext
 	>() =>
 	<Input extends z.ZodObject<any> = any, InferredResponse = unknown, ZodResponse = unknown>({
@@ -228,6 +221,7 @@ const createSubscription =
 			UserRole,
 			CustomClaims,
 			InternalOperationsClient,
+			TypedORM,
 			CustomContext
 		>;
 	} & BaseOperationConfiguration<UserRole>): NodeJSOperation<
@@ -239,6 +233,7 @@ const createSubscription =
 		UserRole,
 		CustomClaims,
 		InternalOperationsClient,
+		TypedORM,
 		CustomContext
 	> => {
 		return {
@@ -256,7 +251,6 @@ const createSubscription =
 			},
 			errors,
 			liveQuery: disabledLiveQueryConfiguration,
-			cache: disabledCacheConfiguration,
 		};
 	};
 
@@ -265,11 +259,12 @@ export const createOperationFactory = <
 	UserRole extends string,
 	CustomClaims extends {},
 	InternalOperationsClient extends OperationsClient,
+	TypedORM,
 	CustomContext
 >() => ({
-	query: createQuery<IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>(),
-	mutation: createMutation<IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>(),
-	subscription: createSubscription<IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>(),
+	query: createQuery<IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>(),
+	mutation: createMutation<IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>(),
+	subscription: createSubscription<IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>(),
 });
 
 export type NodeJSOperation<
@@ -281,6 +276,7 @@ export type NodeJSOperation<
 	UserRole extends string,
 	CustomClaims extends {},
 	InternalOperationsClient extends OperationsClient,
+	TypedORM,
 	CustomContext
 > = {
 	type: OperationType;
@@ -288,17 +284,17 @@ export type NodeJSOperation<
 	responseSchema?: ZodResponse;
 	queryHandler?: ZodResponse extends z.ZodObject<any>
 		? (
-				ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>
+				ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>
 		  ) => Promise<z.infer<ZodResponse>>
 		: (
-				ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>
+				ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>
 		  ) => Promise<Response>;
 	mutationHandler?: ZodResponse extends z.ZodObject<any>
 		? (
-				ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>
+				ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>
 		  ) => Promise<z.infer<ZodResponse>>
 		: (
-				ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, CustomContext>
+				ctx: HandlerContext<Input, IC, UserRole, CustomClaims, InternalOperationsClient, TypedORM, CustomContext>
 		  ) => Promise<Response>;
 	subscriptionHandler?: SubscriptionHandler<
 		Input,
@@ -308,13 +304,14 @@ export type NodeJSOperation<
 		UserRole,
 		CustomClaims,
 		InternalOperationsClient,
+		TypedORM,
 		CustomContext
 	>;
 	errors?: { new (): OperationError }[];
 	requireAuthentication?: boolean;
 	internal: boolean;
 	liveQuery: LiveQueryConfiguration;
-	cache: OperationQueryCacheConfiguration;
+	cache?: QueryCacheConfiguration;
 	rbac: {
 		requireMatchAll: string[];
 		requireMatchAny: string[];
@@ -323,11 +320,14 @@ export type NodeJSOperation<
 	};
 };
 
-export type ExtractInput<B> = B extends NodeJSOperation<infer T, any, any, any, any, any, any, any, any> ? T : never;
+export type ExtractInput<B> = B extends NodeJSOperation<infer T, any, any, any, any, any, any, any, any, any>
+	? T
+	: never;
 export type ExtractResponse<B> = B extends NodeJSOperation<
 	any,
 	infer Response,
 	infer ZodResponse,
+	any,
 	any,
 	any,
 	any,
@@ -342,7 +342,7 @@ export type ExtractResponse<B> = B extends NodeJSOperation<
 
 export const loadNodeJsOperationDefaultModule = async (
 	operationPath: string
-): Promise<NodeJSOperation<any, any, any, any, any, any, any, any, any>> => {
+): Promise<NodeJSOperation<any, any, any, any, any, any, any, any, any, any>> => {
 	// remove .js or / from the end of operationPath if present
 	if (operationPath.endsWith('.cjs')) {
 		operationPath = operationPath.slice(0, -4);