+++ packages/testsuite/apps/basic/.wundergraph/operations/functions/mutation.ts
@@ -1,0 +1,17 @@
+import { createOperation, z } from '../../generated/wundergraph.factory';
+
+export default createOperation.query({
+	input: z.object({
+		name: z.string(),
+	}),
+	handler: async ({ input, operations }) => {
+		const { data } = await operations.mutate({
+			operationName: 'SetName',
+			input: {
+				name: input.name,
+			},
+		});
+
+		return data;
+	},
+});