+++ packages/swr/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.33](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.14.32...@wundergraph/swr@0.14.33) (2023-05-16)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.14.32](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.14.31...@wundergraph/swr@0.14.32) (2023-05-15)
+
+**Note:** Version bump only for package @wundergraph/swr
+
 ## [0.14.31](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.14.30...@wundergraph/swr@0.14.31) (2023-05-12)
 
 **Note:** Version bump only for package @wundergraph/swr